feat(cli): add default proxy support to cli The cli uses the default windows proxy and credentials for requests.
Closes #1373
