+++ framework/src/Volo.Abp.Cli.Core/Volo/Abp/Cli/Http/CliHttpClientHandler.cs
@@ -1,0 +1,14 @@
+﻿using System.Net;
+using System.Net.Http;
+
+namespace Volo.Abp.Cli.Http
+{
+    public class CliHttpClientHandler : HttpClientHandler
+    {
+        public CliHttpClientHandler()
+        {
+            Proxy = WebRequest.GetSystemWebProxy();
+            DefaultProxyCredentials = CredentialCache.DefaultCredentials;
+        }
+    }
+}
